rename index controller
